Added support for widescreen by changing the if statement in main into an if/elseif statement.
This allows two attempts to create the console window and start the program.

If the first if check fails, then it is likely the program is being run on a widescreen monitor, and thus the height of the window was too large. The else if catches this and tries again with a smaller height.