PART 2 OF VIDEO SERIES DONE. Changes in this commit:
Created a function to load in objects to be rendered, replacing the custom cube in earlier versions. .obj files can now be loaded and rendered properly.

Reworked the drawing system to store triangles in a vector, sort them, and then draw them.

added a default .obj file to the program, "VideoShip.obj"